+++ srcjs/columns.js
@@ -426,7 +426,7 @@ export function formatValue(value, options) {
   } = options
 
   if (typeof value === 'number') {
-    if (separators || percent || currency || digits != null) {
+    if (separators || percent || currency || digits != null || locales) {
       // While Number.toLocaleString supports up to 20 fraction digits,
       // IE11 only supports up to 18 digits when formatting as percentages.
       const maximumFractionDigits = 18